Add improved substring implementation that pre-allocates
The benchmark code also adds tests for different implementations of the
substring function. Ideally this would be generic enough that we only
need to add new functions to a list, however I was running into some
issues getting a fully generic version to properly work with the borrow
checker. This item is left as a TODO for now.