Add implementation using rolling adler32 hashes
Properly using the rolling hash (to compute the next hash by removing the
leading character and adding the next character) necessitates using the
raw hash table so we can directly pass in the hash. Unfortunately the
performance I got using the raw hash table is significantly worse than
the performance using HashSet. Profiling shows that the majority of this
slowdown comes from calling the `find()` function on the table. I was
not yet able to determine why this was the case and how to fix it.